pyproject.toml
[project]
name = "technotaggr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "mishkin", email = "mlgelman@usc.edu" }
]
requires-python = ">=3.10"
dependencies = [
    "certifi>=2025.11.12",
    "essentia-tensorflow>=2.1b6.dev1389",
    "gradio>=6.0.1",
    "ipykernel>=7.1.0",
    "librosa>=0.11.0",
    "mutagen>=1.47.0",
    "notebook>=7.5.0",
]

[project.scripts]
technotaggr = "technotaggr:main"
technotaggr-ui = "technotaggr.app:launch_app"

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"
